refactoring in preparation for pretty mode
