+++ packages/wrangler/src/__tests__/tail.test.ts
@@ -1,3 +1,5 @@
+import { DEBUG_MODE_ENABLED } from "../tail";
+
 import { mockAccountId, mockApiToken } from "./helpers/mock-account-id";
 import { setMockResponse } from "./helpers/mock-cfetch";
 import { mockConsoleMethods } from "./helpers/mock-console";
@@ -5,10 +7,6 @@ import { WS } from "./helpers/mock-web-socket";
 import { runInTempDir } from "./helpers/run-in-tmp";
 import { runWrangler } from "./helpers/run-wrangler";
 import type Websocket from "ws";
-
-// change this if you're testing using debug mode
-// (sends all logs regardless of filters)
-const DEBUG = false;
 
 describe("tail", () => {
   runInTempDir();
@@ -18,7 +16,7 @@ describe("tail", () => {
   const std = mockConsoleMethods();
   const api = mockWebsocketApis();
 
-  /* API related functionality */
+  /* Core functionality */
 
   it("creates and then delete tails", async () => {
     expect(api.requests.creation.count).toStrictEqual(0);
@@ -31,6 +29,13 @@ describe("tail", () => {
 
     api.ws.close();
     expect(api.requests.deletion.count).toStrictEqual(1);
+  });
+
+  it("errors if the connection is closed unexpectedly", async () => {
+    api.ws.close();
+    await expect(runWrangler("tail test-worker")).rejects.toThrowError(
+      "Websocket closed unexpectedly!"
+    );
   });
 
   /* filtering */
@@ -142,7 +147,7 @@ describe("tail", () => {
         { client_ip },
         { query },
       ],
-      debug: DEBUG,
+      debug: DEBUG_MODE_ENABLED,
     };
 
     await runWrangler(`tail test-worker ${cliFilters}`);
